requirements.txt
requests
beautifulsoup4
pysolr
python-dotenv
